template-be-flask-main/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
neo4j>=5.18.0
pydantic>=2.9.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
